@@ -1,5 +1,5 @@
 source 'https://rubygems.org'
-ruby '2.5.3'
+ruby '2.6.0'
 
 gem 'sinatra', '~> 2.0'
 gem 'puma', '~> 3.11'
